requirements.txt
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.1
decorator==4.3.0
fuzzywuzzy==0.16.0
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.1.0
pkg-resources==0.0.0
pyparsing==2.2.0
pytesseract==0.2.2
python-dateutil==2.7.3
python-Levenshtein==0.12.0
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
